handle adding new words + hit enter or delete edge case & re-align-mseems to work most of the times, but worth doing some further testing
