+++ package.json
@@ -57,6 +57,7 @@
     "docx": "4.7.1",
     "fontsource-roboto": "^4.0.0",
     "lodash": "^4.17.20",
+    "p-debounce": "^3.0.1",
     "prop-types": "^15.7.2",
     "sbd": "^1.0.18",
     "slate": "^0.59.0",
