+++ src/components/SideBtns/index.js
@@ -1,0 +1,338 @@
+import React, { useState } from 'react';
+import Divider from '@material-ui/core/Divider';
+import Tooltip from '@material-ui/core/Tooltip';
+import Button from '@material-ui/core/Button';
+import Menu from '@material-ui/core/Menu';
+import Link from '@material-ui/core/Link';
+import Grid from '@material-ui/core/Grid';
+import MenuItem from '@material-ui/core/MenuItem';
+import SaveAltIcon from '@material-ui/icons/SaveAlt';
+import SaveOutlinedIcon from '@material-ui/icons/SaveOutlined';
+import KeyboardReturnOutlinedIcon from '@material-ui/icons/KeyboardReturnOutlined';
+import MusicNoteOutlinedIcon from '@material-ui/icons/MusicNoteOutlined';
+import HelpOutlineOutlinedIcon from '@material-ui/icons/HelpOutlineOutlined';
+import ImportExportIcon from '@material-ui/icons/ImportExport';
+import CachedOutlinedIcon from '@material-ui/icons/CachedOutlined';
+import InfoOutlined from '@material-ui/icons/InfoOutlined';
+import PauseOutlinedIcon from '@material-ui/icons/PauseOutlined';
+import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
+import SlateHelpers from '../slate-helpers';
+import subtitlesExportOptionsList from '../../util/export-adapters/subtitles-generator/list.js';
+
+function SideBtns({
+  handleExport,
+  isProcessing,
+  setIsProcessing,
+  insertTextInaudible,
+  handleInsertMusicNote,
+  handleSplitParagraph,
+  isPauseWhiletyping,
+  handleSetPauseWhileTyping,
+  handleRestoreTimecodes,
+  handleReplaceText,
+  handleSave,
+  REPLACE_WHOLE_TEXT_INSTRUCTION,
+}) {
+  const [anchorMenuEl, setAnchorMenuEl] = useState(null);
+  // used by MUI export menu
+  const handleMenuClose = () => {
+    setAnchorMenuEl(null);
+  };
+
+  // used by MUI export menu
+  const handleMenuClick = (event) => {
+    setAnchorMenuEl(event.currentTarget);
+  };
+
+  return (
+    <Grid container direction="row" justify="flex-start" alignItems="flex-start">
+      <div>
+        <Tooltip title={'Export options'}>
+          <Button aria-controls="simple-menu" aria-haspopup="true" onClick={handleMenuClick}>
+            <SaveAltIcon color="primary" /> <KeyboardArrowDownIcon color="primary" />
+          </Button>
+        </Tooltip>
+        <Menu id="simple-menu" anchorEl={anchorMenuEl} keepMounted open={Boolean(anchorMenuEl)} onClose={handleMenuClose}>
+          <MenuItem onClick={handleMenuClose} disabled>
+            <Link style={{ color: 'black' }}>Text Export</Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'text',
+                ext: 'txt',
+                speakers: false,
+                timecodes: false,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary">
+              Text (<code>.txt</code>)
+            </Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'text',
+                ext: 'txt',
+                speakers: true,
+                timecodes: false,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary">Text (Speakers)</Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'text',
+                ext: 'txt',
+                speakers: false,
+                timecodes: true,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary">Text (Timecodes)</Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'text',
+                ext: 'txt',
+                speakers: true,
+                timecodes: true,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary"> Text (Speakers & Timecodes)</Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'text',
+                ext: 'txt',
+                speakers: true,
+                timecodes: true,
+                atlasFormat: true,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary"> Text (Atlas format)</Link>
+          </MenuItem>
+          <Divider />
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'word',
+                ext: 'docx',
+                speakers: false,
+                timecodes: false,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary">
+              {' '}
+              Word (<code>.docx</code>)
+            </Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'word',
+                ext: 'docx',
+                speakers: true,
+                timecodes: false,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary"> Word (Speakers)</Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'word',
+                ext: 'docx',
+                speakers: false,
+                timecodes: true,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary"> Word (Timecodes)</Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'word',
+                ext: 'docx',
+                speakers: true,
+                timecodes: true,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary"> Word (Speakers & Timecodes)</Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'word',
+                ext: 'docx',
+                speakers: false,
+                timecodes: false,
+                inlineTimecodes: true,
+                hideTitle: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary"> Word (OHMS)</Link>
+          </MenuItem>
+          <Divider />
+          <MenuItem onClick={handleMenuClose} disabled>
+            <Link style={{ color: 'black' }}>Closed Captions Export</Link>
+          </MenuItem>
+          {subtitlesExportOptionsList.map(({ type, label, ext }, index) => {
+            return (
+              <MenuItem
+                key={index + label}
+                onClick={() => {
+                  handleExport({ type, ext, isDownload: true });
+                  handleMenuClose();
+                }}
+              >
+                <Link color="primary">
+                  {label} (<code>.{ext}</code>)
+                </Link>
+              </MenuItem>
+            );
+          })}
+          <Divider />
+          <MenuItem onClick={handleMenuClose} disabled>
+            <Link style={{ color: 'black' }}>Developer options</Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'json-slate',
+                ext: 'json',
+                speakers: true,
+                timecodes: true,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary">
+              SlateJs (<code>.json</code>)
+            </Link>
+          </MenuItem>
+          <MenuItem
+            onClick={() => {
+              handleExport({
+                type: 'json-digitalpaperedit',
+                ext: 'json',
+                speakers: true,
+                timecodes: true,
+                isDownload: true,
+              });
+              handleMenuClose();
+            }}
+          >
+            <Link color="primary">
+              DPE (<code>.json</code>)
+            </Link>
+          </MenuItem>
+        </Menu>
+      </div>
+
+      <Tooltip title={'save'}>
+        <Button disabled={isProcessing} onClick={handleSave} color="primary">
+          <SaveOutlinedIcon color="primary" />
+        </Button>
+      </Tooltip>
+
+      {/* TODO: Disabiling until find a way to handle timecodes and alignment on paragraph break */}
+      <Tooltip
+        title={`To insert a paragraph break, and split a pargraph in two, put the cursor at a point where you'd want to add a paragraph break in the text and either click this button or hit enter key`}
+      >
+        <Button disabled={isProcessing} onClick={handleSplitParagraph} color="primary">
+          <KeyboardReturnOutlinedIcon color="primary" />
+        </Button>
+      </Tooltip>
+
+      <Tooltip title={`Put the cursor at a point where you'd want to add [INAUDIBLE] text, and click this button`}>
+        <Button disabled={isProcessing} onClick={insertTextInaudible} color="primary">
+          <HelpOutlineOutlinedIcon color="primary" />
+        </Button>
+      </Tooltip>
+
+      <Tooltip title={'Insert a ♫ in the text'}>
+        <Button disabled={isProcessing} onClick={handleInsertMusicNote} color="primary">
+          <MusicNoteOutlinedIcon color="primary" />
+        </Button>
+      </Tooltip>
+
+      <Tooltip
+        title={` Turn ${isPauseWhiletyping ? 'off' : 'on'} pause while typing functionality. As you start typing the media while pause playback
+                      until you stop. Not reccomended on longer transcript as it might present performance issues.`}
+      >
+        <Button
+          disabled={isProcessing}
+          onClick={handleSetPauseWhileTyping}
+          variant={isPauseWhiletyping ? 'outlined' : null}
+          color={isPauseWhiletyping ? 'secondary' : 'primary'}
+        >
+          <PauseOutlinedIcon color="primary" />
+        </Button>
+      </Tooltip>
+
+      <Tooltip title={' Restore timecodes. At the moment for transcript over 1hour it could temporarily freeze the UI for a few seconds'}>
+        <Button
+          disabled={isProcessing}
+          onClick={async () => {
+            try {
+              setIsProcessing(true);
+              await handleRestoreTimecodes();
+            } finally {
+              setIsProcessing(false);
+            }
+          }}
+          color="primary"
+        >
+          <CachedOutlinedIcon color="primary" />
+        </Button>
+      </Tooltip>
+
+      <Tooltip title={REPLACE_WHOLE_TEXT_INSTRUCTION}>
+        <Button onClick={handleReplaceText} color="primary">
+          <ImportExportIcon color="primary" />
+        </Button>
+      </Tooltip>
+
+      <Tooltip title={' Double click on a word to jump to the corresponding point in the media'}>
+        <Button disabled={isProcessing} color="primary">
+          <InfoOutlined color="primary" />
+        </Button>
+      </Tooltip>
+    </Grid>
+  );
+}
+
+export default SideBtns;
