+++ src/components/index.css
@@ -2,3 +2,28 @@
 .editor-wrapper-container {
   font-family: Roboto, sans-serif;
 }
+
+.editor-wrapper-container {
+  padding: 8px 16px;
+  height: 90vh;
+  overflow: auto;
+}
+/* https://developer.mozilla.org/en-US/docs/Web/CSS/user-select
+TODO: only working in Chrome, not working in Firefox, and Safari - OSX
+if selecting text, not showing selection
+Commented out because it means cannot select speakers and timecode anymore
+which is the intended default behavior but needs to come with export
+functionality to export as plain text, word etc.. otherwise user won't be able
+to get text out of component with timecodes and speaker names in the interim */
+.unselectable {
+  -moz-user-select: none;
+  -webkit-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+}
+.timecode:hover {
+  text-decoration: underline;
+}
+.timecode.text:hover {
+  text-decoration: none;
+}
